pyproject.toml
[tool.poetry]
name = "python-fastapi"
version = "0.1.0"
description = ""
authors = ["Arvind-4 <lordarvindthegamer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
firebase-admin = "^5.2.0"
python-slugify = "^6.1.1"

[tool.poetry.dev-dependencies]
requests = "^2.27.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
